0043 - MarketOrderPaymentCompletedEvent 이벤트 수신 후 주문 품목 불러오기

+++ src/main/java/com/back/boundedContext/market/app/MarketFacade.java
@@ -88,7 +88,7 @@ public class MarketFacade {
     }
 
     @Transactional(readOnly = true)
-    public Optional<Order> findOrderById(int id) {
+    public Optional<Order> findOrderById(long id) {
         return marketSupport.findOrderById(id);
     }
